requirements.txt
asgiref==3.5.2
Django==4.0.5
sqlparse==0.4.2
psycopg2==2.9.3
python-decouple==3.6
gunicorn==20.1.0
djangorestframework==3.13.1
django-cors-headers==3.13.0